Structure of octavIO has been modularized into helpful libraries. Currently, master receives key press data over SPI and sends that over BLE to host MIDI device

+++ MIDI_KEYBOARD_SLAVE1/lib/KeyController/KeyController.h
@@ -37,7 +37,7 @@ public:
     ~KeyController();
 
     void run();
-    void initializeSpi(const uint8_t spiBus, const int spiMode, const size_t bufferSize, const size_t queueSize);
+    void initializeSpi(const uint8_t spiBus, const uint8_t spiMode, const size_t bufferSize, const size_t queueSize);
 
     KeyController() = delete;
     KeyController(const KeyController &) = delete;
